\name{format_nasl}
\alias{format_nasl}
\title{Formats NASL data}
\description{takes data from the Nutrient Analytical Services Lab (NASL) at the
Chesapeake Biological Lab and puts it into a usable dataframe}
\usage{
format_nasl(x)
}
\arguments{
  \item{x}{ Filepath to your .xlsx file in quotes}}

\value{returns a dataframe}
\author{Erin Reilly}
