% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_nasl.R
\name{load_nasl}
\alias{load_nasl}
\title{Load data provided by the Nutrient Analytical Services Lab (NASL) at CBL}
\usage{
load_nasl(data)
}
\arguments{
\item{data}{Filepath the the excel document in quotes.  must be an .xlsx}
}
\value{
A dataframe
}
\description{
This function loads an excel file from NASL. It assumes that the first 23 columns
are superfluous and column names are in the 24th row.
Any rows that are empty will be dropped
}
