% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umol_mg.R
\name{mg_umol}
\alias{mg_umol}
\title{Use the NASL formatted data and convert from mg/L to umol/L}
\usage{
mg_umol(x)
}
\arguments{
\item{data}{dataframe formated using the load_nasl function}
}
\value{
A dataframe with an extra column
}
\description{
This function creates a new column with values in umols/L from the mg/L results
This only works for Nitrogen, Phosphorous, Carbon and Silica in mg/L
All other units will be listed as NA
}
\author{Erin Reilly}

